requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.2
requests==2.32.3
python-dotenv==1.0.1
